+++ samples/attested_tls/CMakeLists.txt
@@ -5,6 +5,7 @@ cmake_minimum_required(VERSION 3.11)
 
 project("Attested TLS sample" LANGUAGES C CXX)
 set(CMAKE_CXX_STANDARD 11)
+set(OE_CRYPTO_LIB mbedtls)
 
 find_package(OpenEnclave CONFIG REQUIRED)
 
@@ -20,18 +21,18 @@ if ((NOT DEFINED ENV{OE_SIMULATION}) OR (NOT $ENV{OE_SIMULATION}))
     COMMENT "Launch processes to establish an Attested TLS between two enclaves"
     COMMAND
       bash -c
-      "${CMAKE_BINARY_DIR}/server/host/tls_server_host ${CMAKE_BINARY_DIR}/server/enc/tls_server_enc.signed -port:12341 &"
+      "${CMAKE_BINARY_DIR}/server/host/tls_server_host ${CMAKE_BINARY_DIR}/server/enc/${OE_CRYPTO_LIB}/tls_server_enc.signed -port:12341 &"
     COMMAND ${CMAKE_COMMAND} -E sleep 2
     COMMAND
       ${CMAKE_BINARY_DIR}/client/host/tls_client_host
-      ${CMAKE_BINARY_DIR}/client/enc/tls_client_enc.signed -server:localhost
-      -port:12341
+      ${CMAKE_BINARY_DIR}/client/enc/${OE_CRYPTO_LIB}/tls_client_enc.signed
+      -server:localhost -port:12341
     COMMAND ${CMAKE_COMMAND} -E sleep 2
     COMMENT
       "Launch processes to establish an Attested TLS between an non-encalve TLS client and an TLS server running inside an enclave "
     COMMAND
       bash -c
-      "${CMAKE_BINARY_DIR}/server/host/tls_server_host ${CMAKE_BINARY_DIR}/server/enc/tls_server_enc.signed -port:12341 &"
+      "${CMAKE_BINARY_DIR}/server/host/tls_server_host ${CMAKE_BINARY_DIR}/server/enc/${OE_CRYPTO_LIB}/tls_server_enc.signed -port:12341 &"
     COMMAND ${CMAKE_COMMAND} -E sleep 2
     COMMAND ${CMAKE_BINARY_DIR}/non_enc_client/tls_non_enc_client
             -server:localhost -port:12341)
