+++ samples/attested_tls/client/Makefile
@@ -4,12 +4,12 @@
 all: build
 
 build:
-	$(MAKE) -C enc
+	$(MAKE) -C enc/${OE_CRYPTO_LIB}
 	$(MAKE) -C host
 
 clean:
-	$(MAKE) -C enc clean
+	$(MAKE) -C enc/${OE_CRYPTO_LIB} clean
 	$(MAKE) -C host clean
 
 run:
-	host/tls_client_host ./enc/tls_client_enclave.signed -server:localhost -port:12341
+	host/tls_client_host ./enc//${OE_CRYPTO_LIB}/tls_client_enclave.signed -server:localhost -port:12341
