+++ samples/attested_tls/client/enc/mbedtls/cert_verifier.cpp
@@ -11,7 +11,7 @@
 #include <mbedtls/platform.h>
 #include <mbedtls/ssl.h>
 #include <string.h>
-#include "../../common/utility.h"
+#include "../../../common/utility.h"
 
 oe_result_t enclave_claims_verifier_callback(
     oe_claim_t* claims,
