+++ samples/attested_tls/client/enc/mbedtls/public.pem
