+++ samples/attested_tls/server/CMakeLists.txt
@@ -1,7 +1,7 @@
 # Copyright (c) Open Enclave SDK contributors.
 # Licensed under the MIT License.
 
-add_subdirectory(enc)
+add_subdirectory(enc/${OE_CRYPTO_LIB})
 add_subdirectory(host)
 
 add_custom_target(tls_server DEPENDS tls_server_host tls_server_enc
