+++ samples/attested_tls/server/Makefile
@@ -4,12 +4,12 @@
 all: build
 
 build:
-	$(MAKE) -C enc
+	$(MAKE) -C enc/${OE_CRYPTO_LIB}
 	$(MAKE) -C host
 
 clean:
-	$(MAKE) -C enc clean
+	$(MAKE) -C enc/${OE_CRYPTO_LIB} clean
 	$(MAKE) -C host clean
 
 run:
-	host/tls_server_host ./enc/tls_server_enc.signed -port:12341
+	host/tls_server_host ./enc/${OE_CRYPTO_LIB}/tls_server_enc.signed -port:12341
