+++ samples/attested_tls/server/enc/mbedtls/CMakeLists.txt
@@ -13,12 +13,13 @@ add_custom_command(
 # Sign enclave and generate a header file for server's MRENCLAVE
 add_custom_command(
   OUTPUT tls_server_enc.signed tls_server_enc_mrenclave.h
-  DEPENDS tls_server_enc enc.conf ${CMAKE_SOURCE_DIR}/server/enc/private.pem
+  DEPENDS tls_server_enc enc.conf
+          ${CMAKE_SOURCE_DIR}/server/enc/${OE_CRYPTO_LIB}/private.pem
           ${CMAKE_SOURCE_DIR}/scripts/gen_mrenclave_header.sh
   COMMAND
     openenclave::oesign sign -e $<TARGET_FILE:tls_server_enc> -c
-    ${CMAKE_SOURCE_DIR}/server/enc/enc.conf -k
-    ${CMAKE_SOURCE_DIR}/server/enc/private.pem
+    ${CMAKE_SOURCE_DIR}/server/enc/${OE_CRYPTO_LIB}/enc.conf -k
+    ${CMAKE_SOURCE_DIR}/server/enc/${OE_CRYPTO_LIB}/private.pem
   COMMAND openenclave::oesign dump -e tls_server_enc.signed > temp.dmp
   COMMAND bash ${CMAKE_SOURCE_DIR}/scripts/gen_mrenclave_header.sh
           ${CMAKE_SOURCE_DIR}/common/tls_server_enc_mrenclave.h temp.dmp
@@ -28,7 +29,7 @@ add_custom_command(
 add_executable(
   tls_server_enc
   ecalls.cpp server.cpp cert_verifier.cpp identity_verifier.cpp
-  ../../common/utility.cpp ${CMAKE_CURRENT_BINARY_DIR}/tls_server_t.c)
+  ../../../common/utility.cpp ${CMAKE_CURRENT_BINARY_DIR}/tls_server_t.c)
 
 if (WIN32)
   maybe_build_using_clangw(tls_server_enc)
@@ -39,7 +40,7 @@ target_compile_definitions(tls_server_enc PUBLIC OE_API_VERSION=2)
 target_include_directories(
   tls_server_enc
   PRIVATE ${CMAKE_CURRENT_SOURCE_DIR} ${CMAKE_CURRENT_BINARY_DIR}
-          ${CMAKE_BINARY_DIR}/server/enc)
+          ${CMAKE_BINARY_DIR}/server/enc/${OE_CRYPTO_LIB})
 
 target_link_libraries(
   tls_server_enc openenclave::oeenclave openenclave::oelibcxx
