+++ samples/attested_tls/server/enc/mbedtls/cert_verifier.cpp
@@ -17,7 +17,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <sys/socket.h>
-#include "../../common/utility.h"
+#include "../../../common/utility.h"
 
 oe_result_t enclave_claims_verifier_callback(
     oe_claim_t* claims,
