+++ samples/attested_tls/server/enc/mbedtls/ecalls.cpp
@@ -2,10 +2,10 @@
 // Licensed under the MIT License.
 
 #include <openenclave/enclave.h>
-#include "../../common/tls_client_enc_pubkey.h"
-#include "tls_client_t.h"
+#include "../../../common/tls_server_enc_pubkey.h"
+#include "tls_server_t.h"
 
-#define ENCLAVE_SECRET_DATA_SIZE 16
+#include <sys/socket.h>
 
 typedef struct _enclave_config_data
 {
@@ -21,19 +21,22 @@ typedef struct _enclave_config_data
 // since the enclave binary itself is not encrypted. Instead, secrets are
 // acquired via provisioning from a service (such as a cloud server) after
 // successful attestation.
-// This g_enclave_secret_data holds the secret data specific to the holding
+// The g_enclave_secret_data holds the secret data specific to the holding
 // enclave, it's only visible inside this secured enclave. Arbitrary enclave
 // specific secret data exchanged by the enclaves. In this sample, the first
 // enclave sends its g_enclave_secret_data (encrypted) to the second enclave.
 // The second enclave decrypts the received data and adds it to its own
 // g_enclave_secret_data, and sends it back to the other enclave.
+#define ENCLAVE_SECRET_DATA_SIZE 16
+
 uint8_t g_enclave_secret_data[ENCLAVE_SECRET_DATA_SIZE] =
     {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16};
 
 enclave_config_data_t config_data = {g_enclave_secret_data,
                                      OTHER_ENCLAVE_PUBLIC_KEY,
                                      sizeof(OTHER_ENCLAVE_PUBLIC_KEY)};
-int ecall_launch_tls_client(char* server_name, char* server_port)
-{
-    return launch_tls_client(server_name, server_port);
-}
+
+// Declare a static dispatcher object for enabling
+// for better organizing enclave-wise global variables
+// static ecall_dispatcher dispatcher("Enclave1", &config_data);
+const char* enclave_name = "Enclave1";
