+++ samples/attested_tls/server/enc/mbedtls/public.pem
