+++ samples/attested_tls/server/enc/mbedtls/server.cpp
@@ -16,7 +16,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <sys/socket.h>
-#include "../../common/utility.h"
+#include "../../../common/utility.h"
 
 extern "C"
 {
